todoAPI/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "todo-api"
version = "1.0.0"
description = "A Flask-based Todo API with JWT authentication"
requires-python = ">=3.11"
authors = [
    {name = "Todo API Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "flask>=2.3.3",
    "flask-sqlalchemy>=3.0.5",
    "flask-jwt-extended>=4.5.3",
    "werkzeug>=2.3.7",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "flask-migrate>=4.0.0",
    "flask-limiter>=3.5.0",
    "flask-cors==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/example/todo-api"
Repository = "https://github.com/example/todo-api"
Issues = "https://github.com/example/todo-api/issues"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
